requirements.txt
asgiref==3.6.0
boto3==1.26.97
botocore==1.29.97
certifi==2022.12.7
Django==4.1.7
python-environ==0.4.54
pywin32==305.1
s3transfer==0.6.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
wincertstore==0.2

